requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.23.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.1

# Model libraries
fairseq>=0.12.2
speechbrain>=0.5.14
pyctcdecode>=0.5.0
kenlm>=0.1

# Data processing
pandas>=2.0.0
pydub>=0.25.1
webrtcvad>=2.0.10
audiomentations>=0.30.0

# Language support
phonemizer>=3.2.1
epitran>=1.23
langdetect>=1.0.9
pypinyin>=0.49.0

# API and serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0

# Evaluation metrics
jiwer>=3.0.0
python-Levenshtein>=0.21.0
sacrebleu>=2.3.1

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development tools
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0
loguru>=0.7.0

# Optional GPU acceleration
# nvidia-ml-py>=11.525.0  # Uncomment for GPU monitoring